refactored to contain all pedal data retrieval and display within FeatureTable component
