+++ src/components/FeatureTable/index.tsx
@@ -1,15 +1,21 @@
 import './index.scss';
-import { useState } from "react";
+import { useState, useEffect } from "react";
+import Loading from '../Loading';
 import FeatureTableCategoryHeader from '../FeatureTableCategoryHeader';
 import FeatureTableColumnHeader from '../FeatureTableColumnHeader';
+import FeatureRow from '../FeatureRow';
+import * as Realm from 'realm-web';
 
-interface Props {
-	children?: React.ReactNode;
-}
+// interface Props {
+// 	children?: React.ReactNode;
+// }
 
-const FeatureTable = ({children}: Props) => {
+const FeatureTable = () => {
 
+	const [pedals, setPedals] = useState([]);
+	const [loading, setLoading] = useState(true);
 	const [tableClasses, setTableClasses] = useState('');
+
 	const categories = [
 		{ label: 'General Info',	clnm: 'general-info-cell',	colHeaders: ['Manufacturer', 'Name', 'Effect Type(s)'] },
 		{ label: 'Signal',			clnm: 'signal-cell',		colHeaders: ['Analog\nSignal', 'True\nBypass', 'Audio\nMix'] },
@@ -24,9 +30,35 @@ const FeatureTable = ({children}: Props) => {
 		{ label: 'Aux Jacks',		clnm: 'aux-jacks-cell',		colHeaders: ['(Hover for\n Details)'] },
 	];
 
+	useEffect(() => {
+		async function getPedalsData () {
+			const realmApp = new Realm.App({id: "pedal_shootout-kycqe"});
+			const creds = Realm.Credentials.anonymous();
+
+			try {
+				const user = await realmApp.logIn(creds);
+				const pedalsArray = await user.functions.listAllPedals();
+				setPedals(pedalsArray);
+			} catch (err) {
+
+			}
+
+			setLoading(false);
+		}
+
+		if (loading) {
+			getPedalsData();
+		}
+	}, [loading])
+
 	return (
 		<div className='feature-table-wrapper'>
 			<h1>Pedals</h1>
+			{loading && (
+				<div className="text-center">
+					<Loading />
+				</div>
+			)}
 			<table id='feature-table' className={tableClasses}>
 				<thead>
 					<tr className='category-header-row'>
@@ -54,7 +86,24 @@ const FeatureTable = ({children}: Props) => {
 					</tr>	
 				</thead>
 				<tbody>
-					{children}
+					{pedals.map((pedal) => {
+						return <FeatureRow
+							key={pedal['pedalId']}
+							pedalManufacturer={pedal['pedalManufacturer']}
+							pedalName={pedal['pedalName']}
+							effectTypes={pedal['effectTypes']}
+							audioSignalType={pedal['audioSignalType']}
+							trueBypass={pedal['isTrueBypassAudioSignal']}
+							audioMix={pedal['audioMix']}
+							hasReorderableLoops={pedal['hasReorderableLoops']}
+							numberOfPresets={pedal['numberOfPresets']}
+							software={pedal['software']}
+							audioConnections={pedal['audioConnections']}
+							powerConnections={pedal['powerConnections']}
+							midiFeatures={pedal['midiFeatures']}
+							auxiliaryJacks={pedal['auxiliaryJacks']}
+						/>
+					})}
 				</tbody>
 			</table>
 		</div>
